+++ src/main/java/web/controller/ReservationController.java
@@ -1,4 +1,4 @@
-package web.presentation;
+package web.controller;
 
 import java.net.URI;
 import java.sql.Date;
